feat: code refactor, added my profile component
